+++ graphql/mutation/ToDo_add.js
@@ -1,38 +1,15 @@
+import chalk from 'chalk';
+
 import { fromGlobalId, mutationWithClientMutationId, cursorForObjectInConnection } from "graphql-relay";
 import { GraphQLString, GraphQLID, GraphQLNonNull } from "graphql";
 
+import { cursorForObjectInConnectionWithUuidComparison } from '../helper/mutation_helper';
 import { DA_User_get } from '../../data/da/User';
 import { DA_ToDo_add, DA_ToDo_get, DA_ToDo_list_get } from '../../data/da/ToDo';
 
 import ToDosConnection from '../type/ToDosConnection';
 import ViewerType from '../type/ViewerType';
 
-import chalk from 'chalk';
-
-function cursorForObjectInConnectionWithUuidComparison( arr, obj )
-{
-  // Make sure that the object and its instance can be compared with ===
-  // assumed that the object has id field which is unique
-  for( let ix = 0; ix < arr.length; ix++ )
-  {
-    if( arr[ ix ].id.equals( obj.id ) )
-    {
-      arr[ ix ] = obj;
-      break;
-    }
-  }
-
-  let cursor = cursorForObjectInConnection( arr, obj );
-  if( cursor == null )
-  {
-    console.log( chalk.bold.red( "Could not create cursor for object in connection" ) );
-    console.log( chalk.gray( "Object: " ) + chalk.red( JSON.stringify( obj ) ) );
-    console.log( chalk.gray( "Array: " ) + chalk.red( JSON.stringify( arr ) ) );
-    console.log( chalk.blue( "." ) );
-  }
-
-  return cursor;
-}
 
 export default mutationWithClientMutationId( {
   name: 'ToDo_add',
